Reward prefix matching
When matching `hea` with `head` and `NoheapProfiling`, it's a bit confusing that both get the same score. 

One way around this is to prime the initial 'current' score in order to reward the first match. If there is a better ways to do this, please point it out.

I'm assuming that this is always desirable, but if not please let me know and I'll add another config parameter